requirements.txt
streamlit
openai
PyMuPDF
Pillow
python-dotenv
tiktoken